powerpc/mm/hash: Convert mask to unsigned long (fate#322495,   bsc#1032471).
suse-commit: 5c473345e4818292ea15a6b327f102a20aaaeab4

+++ arch/powerpc/include/asm/mmu-hash64.h
@@ -318,7 +318,7 @@ static inline unsigned long hpt_hash(unsigned long vpn,
 #endif
 {
 #ifdef CONFIG_BIGMEM
-	int mask;
+	unsigned long mask;
 #endif
 	unsigned long hash, vsid;
